Fixed null pointer exception when managed element is not found
The command_executor has a chance to return an
error code of ENOSUCHELEMENT. This leaves the
element_wrapper object equal to null. Which ends
up throwing an exception on the next like of code
when it tries to use the element. This in turn causes
a crash in IEDriver and it shuts down